Share common code.
